Check document when realtime checking enabled outside of createView
